DataDistributionTester/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
requests

scikit-learn
mediapipe==0.10.0